Some new type fixes for 64bit in our libpurple connector.  Should fix some odd issues, like crashing when initiating an OTR session.

+++ Frameworks/Adium Framework/Source/AIAccount.m
@@ -271,7 +271,7 @@ typedef enum
  *
  * If this implementation is not called, dialog should be released by the subclass.
  */
-- (void)alertForAccountDeletion:(id<AIAccountControllerRemoveConfirmationDialog>)dialog didReturn:(int)returnCode
+- (void)alertForAccountDeletion:(id<AIAccountControllerRemoveConfirmationDialog>)dialog didReturn:(NSInteger)returnCode
 {
 	if(returnCode == NSAlertDefaultReturn) {
 		[self performDelete];
